mv geolocation process into useGeolocation
