+++ src/libs/location/LocationWrapper.tsx
@@ -1,24 +1,13 @@
-import React, { memo, useCallback, useContext, useEffect, useMemo, useRef } from 'react'
-import { Linking } from 'react-native'
+import React, { memo, useContext, useEffect, useMemo, useRef } from 'react'
 
 import { analytics } from 'libs/analytics'
-import { useAppStateChange } from 'libs/appState'
 import { useSafeState } from 'libs/hooks'
-import { checkGeolocPermission } from 'libs/location/geolocation/checkGeolocPermission/checkGeolocPermission'
-import { getGeolocPosition } from 'libs/location/geolocation/getGeolocPosition/getGeolocPosition'
-import { requestGeolocPermission } from 'libs/location/geolocation/requestGeolocPermission/requestGeolocPermission'
+import { useGeolocation } from 'libs/location/geolocation/hook/useGeolocation'
 import { SuggestedPlace } from 'libs/place'
 import { storage } from 'libs/storage'
-import { useModal } from 'ui/components/modals/useModal'
 
 import { GeolocationActivationModal } from './geolocation/components/GeolocationActivationModal'
-import { GeolocPermissionState } from './geolocation/enums'
-import {
-  GeolocationError,
-  ILocationContext,
-  Position,
-  RequestGeolocPermissionParams,
-} from './types'
+import { ILocationContext } from './types'
 
 /* eslint-disable @typescript-eslint/no-empty-function */
 const LocationContext = React.createContext<ILocationContext>({
@@ -41,98 +30,21 @@ export const LocationWrapper = memo(function LocationWrapper({
 }: {
   children: React.JSX.Element
 }) {
-  const [geolocPosition, setGeolocPosition] = useSafeState<Position>(undefined)
   const [place, setPlace] = useSafeState<SuggestedPlace | null>(null)
-  const [geolocPositionError, setGeolocPositionError] = useSafeState<GeolocationError | null>(null)
-  const [permissionState, setPermissionState] = useSafeState<GeolocPermissionState | undefined>(
-    undefined
-  )
-  const hasGeolocPosition = !!geolocPosition
-  const onModalHideRef = useRef<() => void>()
-
   const {
-    visible: isGeolocPermissionModalVisible,
-    showModal: showGeolocPermissionModal,
-    hideModal: hideGeolocPermissionModal,
-  } = useModal(false)
+    geolocPosition,
+    geolocPositionError,
+    permissionState,
+    hasGeolocPosition,
+    triggerPositionUpdate,
+    onPressGeolocPermissionModalButton,
+    isGeolocPermissionModalVisible,
+    showGeolocPermissionModal,
+    hideGeolocPermissionModal,
+    contextualRequestGeolocPermission,
+  } = useGeolocation()
 
-  useEffect(() => {
-    contextualCheckPermission()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  const triggerPositionUpdate = useCallback(async () => {
-    try {
-      const newPosition = await getGeolocPosition()
-      setGeolocPosition(newPosition)
-      setGeolocPositionError(null)
-      return newPosition
-    } catch (e) {
-      const newPositionError = e as { cause: GeolocationError } | null
-      setGeolocPosition(null)
-      setGeolocPositionError(newPositionError?.cause ?? null)
-      return null
-    }
-  }, [setGeolocPosition, setGeolocPositionError])
-
-  // this function is used to set OS permissions according to user choice on native geolocation popup
-  const contextualRequestGeolocPermission = useCallback(
-    async function (params?: RequestGeolocPermissionParams) {
-      !!params?.onSubmit && params.onSubmit()
-
-      let permission = await requestGeolocPermission()
-      if (shouldTriggerPositionUpdate(permission)) {
-        const newPosition = await triggerPositionUpdate()
-        if (isNeedAskPosition(permission)) {
-          permission =
-            newPosition === null
-              ? GeolocPermissionState.NEVER_ASK_AGAIN
-              : GeolocPermissionState.GRANTED
-        }
-      }
-
-      if (isGranted(permission)) {
-        !!params?.onAcceptance && params.onAcceptance()
-      } else {
-        !!params?.onRefusal && params.onRefusal()
-      }
-      setPermissionState(permission)
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    []
-  )
-
-  // in case user updates his preferences in his phone settings we check if his local
-  // storage choice is consistent with phone permissions, and update position if not
-  const contextualCheckPermission = useCallback(async function () {
-    let permission = await checkGeolocPermission()
-    if (shouldTriggerPositionUpdate(permission)) {
-      const newPosition = await triggerPositionUpdate()
-      if (isNeedAskPosition(permission)) {
-        permission =
-          newPosition === null
-            ? GeolocPermissionState.NEVER_ASK_AGAIN
-            : GeolocPermissionState.GRANTED
-      }
-    }
-    setPermissionState(permission)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
-  useEffect(() => {
-    if (isRejected(permissionState)) {
-      setGeolocPosition(null)
-      setGeolocPositionError(null)
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [permissionState])
-
-  useAppStateChange(contextualCheckPermission, undefined, [])
-
-  const onPressGeolocPermissionModalButton = useCallback(() => {
-    Linking.openSettings()
-    hideGeolocPermissionModal()
-  }, [hideGeolocPermissionModal])
+  const onModalHideRef = useRef<() => void>()
 
   useEffect(() => {
     switch (true) {
@@ -190,21 +102,4 @@ export const LocationWrapper = memo(function LocationWrapper({
 
 export function useLocation(): ILocationContext {
   return useContext(LocationContext)
-}
-
-function isGranted(permission: GeolocPermissionState) {
-  return permission === GeolocPermissionState.GRANTED
-}
-function isRejected(permission: GeolocPermissionState | undefined) {
-  return (
-    !permission ||
-    permission === GeolocPermissionState.DENIED ||
-    permission === GeolocPermissionState.NEVER_ASK_AGAIN
-  )
-}
-function isNeedAskPosition(permission: GeolocPermissionState) {
-  return permission === GeolocPermissionState.NEED_ASK_POSITION_DIRECTLY
-}
-function shouldTriggerPositionUpdate(permission: GeolocPermissionState) {
-  return isGranted(permission) || isNeedAskPosition(permission)
 }
